+++ server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/managers/BazelBspCompilationManager.java
@@ -54,6 +54,7 @@ public class BazelBspCompilationManager {
       return new BepBuildResult(result, bepServer.getBepOutput());
     } finally {
       nettyServer.shutdown();
+      executor.shutdown();
     }
   }