+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ## [Unreleased]
 
+### Fixes 🛠️
+- Publish `build/publishDiagnostics` with an empty array of diagnostics to clear former diagnotics.
+  | [#381](https://github.com/JetBrains/bazel-bsp/pull/381)
+
 ## [2.7.0]
 
 ### Features 🎉
@@ -16,8 +20,6 @@
   | [#372](https://github.com/JetBrains/bazel-bsp/pull/372)
 - Set PublishDiagnosticsParams.reset to be true.
   | [#377](https://github.com/JetBrains/bazel-bsp/pull/377)
-- Publish `build/publishDiagnostics` with an empty array of diagnostics to clear former diagnotics.
-  | [#381](https://github.com/JetBrains/bazel-bsp/pull/381)
 - Update document about how to use projectview.
   | [#383](https://github.com/JetBrains/bazel-bsp/pull/383)
 - Fixup failed target names from BEP in bazel 6+ in the bloop mode.