+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 
 ## [Unreleased]
+### Fixes 🛠️
+-  Prioritize most frequently used JDKs when selecting the project JDK.
+   | [#420](https://github.com/JetBrains/bazel-bsp/pull/420)
+-  Fix ide classpath computation for recent rules_jvm_external.
+   | [#421](https://github.com/JetBrains/bazel-bsp/pull/421)
+-  Process exit instead of hang in case of uncaught exception in pooled threads
+   | [#425](https://github.com/JetBrains/bazel-bsp/pull/425)
 
 ### Fixes 🛠️
 - Publish `build/publishDiagnostics` with an empty array of diagnostics to clear former diagnotics.